pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voiceflow"
version = "0.1.0"
description = "Voice companion — ambient voice shell with TTS, wake-word, and WebSocket daemon"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "websockets>=12.0",
    "pyttsx3>=2.90",
    "pyyaml>=6.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
listen = [
    "vosk>=0.3.45",
    "sounddevice>=0.4.6",
    "pvporcupine>=3.0.0",
    "pvrecorder>=1.2.0",
]
ui = [
    "PyQt6>=6.5.0",
    "pygame>=2.5.0",
]
actions = [
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "pygetwindow>=0.0.9",
]
full = ["voiceflow[listen,ui,actions]"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
vf = "voiceflow.cli:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src/voiceflow"]
